Added the code for the friends api endpoint
Now you can get a list of friends through /api/friends in the JSON format. To make this work I added that accessToken for the graph API is saved to the server when the user signs up. I am not sure yet how often the accessToken needs to be renewed (I think it expires after 2 hours), but maybe we can figure that out later.