Strange behavior with the API GET (/api/friends) and PUT (/api/users/:fbid)
Currently, when the server is started, and you log in as user, the PUT request to api/users/:fbid doesn't work, but the GET request to api/friends works.

Then, if you restart the server, WITHOUT refreshing the page, the PUT request works, but the GET request gets response 302 "moved temporarily"

Somehow the requests must be changing, but they look the same?