The put request now works for me
The problem with the put request for me was that it was using the default object from backbone - which contained your fbId and not mine. I know changed it to the way it is used in api/friends, with not using the params at all, but just using the user object coming with the request (I don't know where it comes from, maybe from the cookie?).

This also means that the fbId that backbone is giving the put request is actually not needed at all (and also wrong), but at least it is now updating the logged in user and not the default one from backbone