Added MongoDB script for test users
I added a script for populating the database with two test users. You can run it through 'mongo mongo_fill.js'. It actually clears the whole collection before, to make sure that there is no wrong data in it (That can break things..) . I used Pia and Oleg, because they are shared friends of ours. Also while testing it I noticed that the deserialize function worked wrongly. It always took the first user in the database not the actual user. I only noticed it because I put in the test users first and then signed in.