Changed the views and templates
I had a look at the views, because I thought about how to add the configuration views. I kind of thought that we right now have to many views, therefore I merged the NotUpFo and IsUpFo into UpFoView. It still works and makes more sense for me. Also the list of friends to invite is now always shown