Refactor geocode and geocode when changing options
Put the geocode functions to the helpers. How it's also used in the
options page. When the name is changed, the location is geocoded again.
What's missing is that if the location would be changed in the profile,
it should only be possible to save it, if the location is geocodable.