Reverse geocoding for the (browser) location
Now when a new lat/lng location is saved, automatically also the city
is updated like "Helsinki, FI" for example.

Also I commented out some of the console.logs that were printing all
the facebook friends.

Right now the reverse geocoding is not in use yet, but I tested it with
the .browserLocation function on the front-end.