Refactor the location 'sync'
I refactored the location stuff, now it is a little bit easier and the
callbacks work correctly. The options page is only still missing
something to prompt that the updated city was not found. But I think
the city selection should not be free text anyway - but like in
facebook - with autocomplete.